% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_boot.R
\name{double_boot}
\alias{double_boot}
\title{Double bootstrap for selecting alpha to use with the m-out-of-n bootstrap}
\usage{
double_boot(i, eta, estVal, B2, see_df_list, coef_dr, current_alpha)
}
\arguments{
\item{i}{Scalar representing a seed or iteration}

\item{eta}{Scalar for the (1-eta)% confidence set}

\item{estVal}{Scalar for the estimate of Value that the confidence set is meant to surround}

\item{B2}{Scalar for number of conditional bootstraps to run}

\item{see_df_list}{List of imputed data sets}

\item{coef_dr}{Vector representing decision rule components in the form
of c(intercept, trtmt, main effects, main effects by trtmt interactions)}

\item{current_alpha}{Scalar for value of alpha used}
}
\value{
Scalar indicator for whether the double centered percentile bootstrap captures the estimate of Value
}
\description{
Double bootstrap for selecting alpha to use with the m-out-of-n bootstrap
}
