% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genDat.R
\name{genDat}
\alias{genDat}
\title{Generate a simulated data set containing missingness}
\usage{
genDat(
  size,
  propMiss,
  seedNum,
  settingMiss = "Original",
  settingNumber,
  splitData = FALSE
)
}
\arguments{
\item{size}{Size of simulated data set}

\item{propMiss}{Proportion of missingness (30% or 50%)}

\item{seedNum}{Seed used for random generation}

\item{settingMiss}{Missingness setting, which could be Original or Modified (to be added later)}
}
\value{
List of data frames containing the full data set and the data set with missingness
}
\description{
Generate a simulated data set containing missingness
}
