% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_m_alpha.R
\name{get_m_alpha}
\alias{get_m_alpha}
\title{Estimate of m for the m-out-of-n bootstrap for a given value of alpha}
\usage{
get_m_alpha(see_df, alpha, coef_dr)
}
\arguments{
\item{see_df}{Data frame of interest}

\item{alpha}{Scalar for the alpha value used in estimating m for the m-out-of-n bootstrap}
}
\value{
Scalar estimate of m
}
\description{
Estimate of m for the m-out-of-n bootstrap for a given value of alpha
}
