% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mBoot_CS.R
\name{mboot_CS}
\alias{mboot_CS}
\title{Confidence Set of m-out-of-n bootstrap results}
\usage{
mboot_CS(eta, m, mboot_multi_res, estVal)
}
\arguments{
\item{eta}{Scalar for the (1-eta)\% confidence set}

\item{m}{Scalar for subsample size, where m < n}

\item{mboot_multi_res}{Vector of results from multiple repetitions of the m-out-of-n bootstrap}

\item{estVal}{Scalar for estimate of Value that the confidence set surrounds}
}
\value{
Vector representing the lower and upper bounds of the (1-eta)\% confidence set
}
\description{
Confidence Set of m-out-of-n bootstrap results
}
