% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mBoot_Val.R
\name{mboot_Val}
\alias{mboot_Val}
\title{Estimate AIPW estimate of Value for a subsample of size m on a given data set}
\usage{
mboot_Val(m, see_df, coef_dr)
}
\arguments{
\item{m}{Scalar for subsample size, where m < n}

\item{see_df}{Data set used for estimating Value}

\item{coef_dr}{Vector representing decision rule components in the form
of c(intercept, trtmt, main effects, main effects by trtmt interactions)}
}
\value{
Scalar AIPW estimate of Value on the subsample of size m
}
\description{
Estimate AIPW estimate of Value for a subsample of size m on a given data set
}
