% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mBoot_Val_multi.R
\name{mboot_Val_multi}
\alias{mboot_Val_multi}
\title{Perform multiple repetitions of estimating Value on different subsamples of size m}
\usage{
mboot_Val_multi(m, see_df, coef_dr, reps)
}
\arguments{
\item{m}{Scalar for subsample size, where m < n}

\item{see_df}{Data set used for estimating Value}

\item{coef_dr}{Vector representing decision rule components in the form
of c(intercept, trtmt, main effects, main effects by trtmt interactions)}

\item{reps}{Scalar for number of repetitions to perform}
}
\value{
Vector of estimates of Value (AIPW) on different subsamples of size m
}
\description{
Perform multiple repetitions of estimating Value on different subsamples of size m
}
