% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mn_double_Val.R
\name{mn_double_Val}
\alias{mn_double_Val}
\title{Run the double bootstrap procedure for selecting alpha to use with the m-out-of-n bootstrap}
\usage{
mn_double_Val(missDat, numImp, alpha_see, seedNum, B1, B2, eta = 0.05, nCores)
}
\arguments{
\item{missDat}{Data frame containing missingness}

\item{numImp}{Scalar for number of imputations}

\item{alpha_see}{Vector of candidate alpha values}

\item{seedNum}{Scalar for seed}

\item{B1}{Scalar for number of outer bootstraps to run}

\item{B2}{Scalar for number of conditional bootstraps to perform}

\item{eta}{Scalar for the (1-eta)% confidence set}

\item{nCores}{If "LSB", function is run using the number of cores specified for the LPC grid;
Otherwise, a user-specified number of cores to use}
}
\value{
Vector for the recommended alpha and the corresponding coverage
}
\description{
Run the double bootstrap procedure for selecting alpha to use with the m-out-of-n bootstrap
}
