% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mn_Val.R
\name{mn_Val}
\alias{mn_Val}
\title{Run the m-out-of-n bootstrap}
\usage{
mn_Val(missDat, numImp, seedNum, reps, eta = 0.05, alpha = 0.025)
}
\arguments{
\item{missDat}{Data frame containing missingness}

\item{numImp}{Scalar for number of imputations}

\item{seedNum}{Scalar for seed}

\item{reps}{Scalar for number of m-out-of-n bootstrap repetitions to perform}

\item{eta}{Scalar for the (1-eta)% confidence set}

\item{alpha}{Scalar for the alpha value to use when estimating m}
}
\value{
List containing the 95% CI formed with the m-out-of-n bootstraps from R imputations,
the choice of 'm', and the final 95% CI with m-out-of-n bootstrap
}
\description{
Run the m-out-of-n bootstrap
}
