% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recTrt_dr.R
\name{recTrt_dr}
\alias{recTrt_dr}
\title{Return vector of recommended treatments under decision rule}
\usage{
recTrt_dr(matCov, coef_dr)
}
\arguments{
\item{matCov}{Matrix of covariate information}

\item{coef_dr}{Vector representing decision rule components in the form
of c(intercept, trtmt, main effects, main effects by trtmt interactions)}
}
\value{
Vector of recommended treatments
}
\description{
Return vector of recommended treatments under decision rule
}
