% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_double_boot.R
\name{run_double_boot}
\alias{run_double_boot}
\title{Run the double bootstrap procedure for selecting alpha to use with the m-out-of-n bootstrap}
\usage{
run_double_boot(
  missDat,
  numImp,
  alpha_see,
  seedNum,
  iters,
  B2,
  eta = 0.05,
  runLPC
)
}
\arguments{
\item{missDat}{Data frame containing missingness}

\item{numImp}{Scalar for number of imputations}

\item{alpha_see}{Vector of candidate alpha values}

\item{seedNum}{Scalar for seed}

\item{iters}{Scalar for number of iterations to perform for outer bootstrap}

\item{B2}{Scalar for number of conditional bootstraps to perform}

\item{eta}{Scalar for the (1-eta)\% confidence set}

\item{runLPC}{If TRUE, function is run using the number of cores specified for the LPC grid;
Otherwise, a user-specified number of cores to use}
}
\value{
Vector for the recommended alpha and the corresponding coverage
}
\description{
Run the double bootstrap procedure for selecting alpha to use with the m-out-of-n bootstrap
}
