% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mBoot.R
\name{run_mBoot}
\alias{run_mBoot}
\title{Run the m-out-of-n bootstrap}
\usage{
run_mBoot(missDat, numImp, seedNum, reps, eta = 0.05)
}
\arguments{
\item{missDat}{Data frame containing missingness}

\item{numImp}{Scalar for number of imputations}

\item{seedNum}{Scalar for seed}

\item{reps}{Scalar for number of m-out-of-n bootstrap repetitions to perform}

\item{eta}{Scalar for the (1-eta)\% confidence set}
}
\value{
List containing the 95\% CI formed with the m-out-of-n bootstraps from R imputations,
the choice of 'm', and the final 95\% CI with m-out-of-n bootstrap
}
\description{
Run the m-out-of-n bootstrap
}
