% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_Val.R
\name{split_Val}
\alias{split_Val}
\title{Estimate Value using the data splitting approach on data sets with missingness}
\usage{
split_Val(missDat, numImp, seedNum)
}
\arguments{
\item{missDat}{Data set containing missingness}

\item{numImp}{Number of imputations for imputing with MICE}
}
\value{
Estimates of the Value and variance of the Value under the IPW and AIPW estimators
}
\description{
Estimate Value using the data splitting approach on data sets with missingness
}
