% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValEst.R
\name{ValEst}
\alias{ValEst}
\title{Estimate the Value and variance of the Value using the AIPW and IPW estimators
Based on the formulas provided in the book, "Dynamic Treatment Regimes" (2020) by Tsiatis, Davidian, Holloway, and Laber}
\usage{
ValEst(testDf, coef_dr, isSplit = TRUE, estVar = TRUE)
}
\arguments{
\item{testDf}{Data frame of interest for estimating the Value and variance of the Value}

\item{coef_dr}{Vector representing decision rule components in the form
of c(intercept, trtmt, main effects, main effects by trtmt interactions)}

\item{isSplit}{Boolean for whether the Value estimate is obtained on a split data set (rather than full)}

\item{estVar}{If TRUE, will also estimate the variance of the Value. Otherwise, will just return the AIPW estimate of the Value.}
}
\value{
Either a scalar representing the AIPW estimate of Value or
A vector of the estimate of Value and the variance of the Value for the IPW and AIPW estimators
}
\description{
Estimate the Value and variance of the Value using the AIPW and IPW estimators
Based on the formulas provided in the book, "Dynamic Treatment Regimes" (2020) by Tsiatis, Davidian, Holloway, and Laber
}
